agentic_ai_project/requirements.txt
google-generativeai
python-dotenv
pyyaml
pydantic
pandas
numpy
structlog
opentelemetry-api
opentelemetry-sdk
fastapi
uvicorn
